Icon is loaded only once with the map
